pyproject.toml
[build-system]
requires = ["setuptools>=68", "setuptools-scm>=8"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "issue-observatory"
version = "0.1.0"
description = "A platform for systematic collection and analysis of public discourse across digital arenas"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }

dependencies = [
    # Web framework
    "fastapi>=0.115,<0.116",
    "uvicorn[standard]>=0.32,<0.33",
    "jinja2>=3.1,<4.0",
    "python-multipart>=0.0.12,<0.1",
    # Database
    "sqlalchemy[asyncio]>=2.0,<3.0",
    "asyncpg>=0.30,<0.31",
    "alembic>=1.14,<2.0",
    # Validation & settings
    "pydantic>=2.10,<3.0",
    "pydantic-settings>=2.7,<3.0",
    "pydantic[email]>=2.10,<3.0",
    # Auth
    "fastapi-users[sqlalchemy]>=13.0,<14.0",
    "python-jose[cryptography]>=3.3,<4.0",
    "passlib[bcrypt]>=1.7,<2.0",
    "cryptography>=44,<45",
    # Task queue
    "celery[redis]>=5.4,<6.0",
    "redis>=5.2,<6.0",
    # HTTP
    "httpx>=0.28,<0.29",
    # Feed parsing
    "feedparser>=6.0,<7.0",
    # Reddit async API wrapper
    "asyncpraw>=7.7,<8.0",
    # WebSocket support (Bluesky Jetstream streamer)
    "websockets>=13.0,<14.0",
    # Telegram MTProto client (Telegram arena)
    "telethon>=1.36,<2.0",
    # Logging
    "structlog>=24.4,<25.0",
    # Object storage (MinIO / S3-compatible) — used by backup scripts
    "minio>=7.2,<8.0",
]

[project.optional-dependencies]
ml = [
    # Optional dependency for content-similarity scoring in sampling/similarity_finder.py.
    # When not installed, find_similar_by_content() falls back to Jaccard word overlap.
    "scikit-learn>=1.4,<2.0",
]
dev = [
    "pytest>=8.3,<9.0",
    "pytest-asyncio>=0.24,<0.25",
    "pytest-cov>=6.0,<7.0",
    "ruff>=0.8,<0.9",
    "mypy>=1.13,<2.0",
    "pre-commit>=4.0,<5.0",
    "factory-boy>=3.3,<4.0",
    "httpx>=0.28,<0.29",
    "respx>=0.21,<0.22",
    # Type stubs
    "types-passlib>=1.7,<2.0",
    "types-redis>=4.6,<5.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

# ---------------------------------------------------------------------------
# Ruff
# ---------------------------------------------------------------------------
[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ANN", # flake8-annotations
    "TCH", # flake8-type-checking
    "RUF", # ruff-specific rules
]
ignore = [
    "ANN101",  # Missing type annotation for `self`
    "ANN102",  # Missing type annotation for `cls`
    "ANN401",  # Dynamically typed expressions (Any)
    "B008",    # Do not perform function calls in argument defaults (FastAPI depends)
]

[tool.ruff.lint.isort]
known-first-party = ["issue_observatory"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

# ---------------------------------------------------------------------------
# Mypy
# ---------------------------------------------------------------------------
[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["feedparser.*", "jose.*", "passlib.*", "celery.*", "factory.*"]
ignore_missing_imports = true

# ---------------------------------------------------------------------------
# Pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = [
    "--cov=src/issue_observatory",
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov",
    "-v",
]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]
